Added the p:encode step
Fix #659

In order to make the p:cast-content-type step decode any c:data document, even
one that might contain serialized XML, I’ve reorganized the “Casting from an XML
media type” section. The case of a c:data document is handled first.

This is a small backwards incompatibility because it means that you cannot use
p:cast-content type to change the content type of a c:data element from
application/xml to, for example, application/foo+xml. But that seems like
something very unlikely given that we control the semantics of c:data
documents. (The alternative is to create a p:decode step.)